Split up the policies on two groups to prevent reaching the quota
In https://github.com/stratospheric-dev/stratospheric/issues/191#issuecomment-1441596512
I was told to split up the policies into several groups which is
obviously the right solution. I just didn't see it before.

I was too lazy for now to come up with good group names for the CI user
and just split them up and did not look out for good semantic groupings.
Maybe later, when I found a solution for the problem, that the CI user
can not assume the role with the right to deploy.